added cli api (new methods) to test-env-js and did minor refactoring
