+++ packages/js/test-env/src/test-environment.ts
@@ -1,0 +1,76 @@
+import { awaitResponse } from "./util";
+import { ensAddresses, runCLI } from "./index";
+
+export const initTestEnvironment = async (cli?: string): Promise<void> => {
+  // Start the test environment
+  const { exitCode, stderr, stdout } = await runCLI({
+    args: ["infra", "up", "--modules=eth-ens-ipfs", "--verbose"],
+    cli,
+  });
+
+  if (exitCode) {
+    throw Error(
+      `initTestEnvironment failed to start test environment.\nExit Code: ${exitCode}\nStdErr: ${stderr}\nStdOut: ${stdout}`
+    );
+  }
+
+  // Wait for all endpoints to become available
+  let success = false;
+
+  // IPFS
+  success = await awaitResponse(
+    `http://localhost:5001/api/v0/version`,
+    '"Version":',
+    "get",
+    2000,
+    20000
+  );
+
+  if (!success) {
+    throw Error("test-env: IPFS failed to start");
+  }
+
+  // Ganache
+  success = await awaitResponse(
+    `http://localhost:8545`,
+    '"jsonrpc":',
+    "post",
+    2000,
+    20000,
+    '{"jsonrpc":"2.0","method":"eth_blockNumber","params":[],"id":83}'
+  );
+
+  if (!success) {
+    throw Error("test-env: Ganache failed to start");
+  }
+
+  // ENS
+  success = await awaitResponse(
+    "http://localhost:8545",
+    '"result":"0x',
+    "post",
+    2000,
+    20000,
+    `{"jsonrpc":"2.0","method":"eth_getCode","params":["${ensAddresses.ensAddress}", "0x2"],"id":1}`
+  );
+
+  if (!success) {
+    throw Error("test-env: ENS failed to deploy");
+  }
+};
+
+export const stopTestEnvironment = async (cli?: string): Promise<void> => {
+  // Stop the test environment
+  const { exitCode, stderr } = await runCLI({
+    args: ["infra", "down", "--modules=eth-ens-ipfs"],
+    cli,
+  });
+
+  if (exitCode) {
+    throw Error(
+      `stopTestEnvironment failed to stop test environment.\nExit Code: ${exitCode}\nStdErr: ${stderr}`
+    );
+  }
+
+  return Promise.resolve();
+};
