+++ packages/js/test-env/src/util.ts
@@ -1,0 +1,63 @@
+import axios from "axios";
+
+export async function awaitResponse(
+  url: string,
+  expectedRes: string,
+  getPost: "get" | "post",
+  timeout: number,
+  maxTimeout: number,
+  data?: string
+): Promise<boolean> {
+  let time = 0;
+
+  while (time < maxTimeout) {
+    const request = getPost === "get" ? axios.get(url) : axios.post(url, data);
+    const success = await request
+      .then(function (response) {
+        const responseData = JSON.stringify(response.data);
+        return responseData.indexOf(expectedRes) > -1;
+      })
+      .catch(function () {
+        return false;
+      });
+
+    if (success) {
+      return true;
+    }
+
+    await new Promise<void>(function (resolve) {
+      setTimeout(() => resolve(), timeout);
+    });
+
+    time += timeout;
+  }
+
+  return false;
+}
+
+export async function awaitPing(
+  url: string,
+  timeout: number,
+  maxTimeout: number
+): Promise<boolean> {
+  let time = 0;
+
+  while (time < maxTimeout) {
+    const request = axios.get(url, { timeout });
+    const success = await request
+      .then(() => true)
+      .catch((e) => e.code !== "ECONNRESET");
+
+    if (success) {
+      return true;
+    }
+
+    await new Promise<void>(function (resolve) {
+      setTimeout(() => resolve(), timeout);
+    });
+
+    time += timeout;
+  }
+
+  return false;
+}
